feat(storage): update request models for appendable upload
